schedule redraw better
Previously the code was just waiting whatever number of seconds per
year/month/day/minute, depending on what the most human readable unit
happens to be at the time, but if the redraw happens at some time other
than when the text needs to be updated, then the next redraw time needs
to be shorter than the whole time period.

Subtract from the time period the amount of time into the time period
which has already passed.  That will leave the time left to the next
time the text should be updated.